Limiting allowed difficulty values
If difficulty is too large then the first number in a division problem
can be oversized which will mess up the layout.  There isn't much point
to drilling with large numbers anyway, so the value is limited to 99.

Also need to prevent negative numbers and numbers smaller than 10 don't
make sense either.

So the difficulty is allowed to be (10-99).