Fraction options updated
The settings for proper/improper fractions are on the menu.  The
fraction-specific setting appears even on the integers activity, which
is not ideal.  Also, the appearance of the settings menu changes between
the activities and there are multiple different layouts for the fraction
activities and I'm not sure what some of them do (I think it might be
for the initial loading prior to requesting landscape mode, maybe?).

The update is ready to push; everything else can be v2.1