Fixed issue with negatives
Object equality is a stringent condition and so (-1/2) is not equal to
(1/-2).  After allowing negatives to be generated in the problems, the
situation could arise where the correct answer might return false due to
this inequality.  To prevent this a "canonical form" is defined with at
most one negative, always appearing in the numerator.